Move mappers and tests to new module eclipsestore-processor
