--- eclipsestore-annotations/src/main/resources/META-INF/services/io.micronaut.inject.annotation.AnnotationMapper
@@ -1,3 +1,0 @@
-io.micronaut.eclipsestore.annotations.StoreRootAnnotationMapper
-io.micronaut.eclipsestore.annotations.StoreReturnAnnotationMapper
-io.micronaut.eclipsestore.annotations.StoreParamsAnnotationMapper
