+++ eclipsestore-processor/src/main/java/io/micronaut/eclipsestore/processor/StoreParamsAnnotationMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2022 original authors
+ * Copyright 2017-2024 original authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.eclipsestore.annotations;
+package io.micronaut.eclipsestore.processor;
 
 import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Internal;
+import io.micronaut.eclipsestore.annotations.Store;
+import io.micronaut.eclipsestore.annotations.StoreParams;
 import io.micronaut.inject.annotation.TypedAnnotationMapper;
 import io.micronaut.inject.visitor.VisitorContext;
 
 import java.util.List;
-
 
 /**
  * Maps the {@link StoreParams} annotation to the {@link Store} annotation.
