+++ eclipsestore-processor/src/main/resources/META-INF/services/io.micronaut.inject.annotation.AnnotationMapper
@@ -1,0 +1,3 @@
+io.micronaut.eclipsestore.processor.StoreRootAnnotationMapper
+io.micronaut.eclipsestore.processor.StoreReturnAnnotationMapper
+io.micronaut.eclipsestore.processor.StoreParamsAnnotationMapper
